requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
discord.py==2.3.2
python-dotenv==1.0.0
chromadb==0.4.22
sentence-transformers==2.3.1
numpy<2.0
torch
torchvision
torchaudio
groq